+++ packages/orq-rc/RELEASES.md
@@ -9519,3 +9519,13 @@ Based on:
 - [python v4.12.0-rc.31] packages/orq-rc
 ### Releases
 - [PyPI v4.12.0-rc.31] https://pypi.org/project/orq-ai-sdk/4.12.0-rc.31 - packages/orq-rc
+
+## 2026-07-16 07:25:32
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.790.2 (2.918.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.12.0-rc.32] packages/orq-rc
+### Releases
+- [PyPI v4.12.0-rc.32] https://pypi.org/project/orq-ai-sdk/4.12.0-rc.32 - packages/orq-rc